extract connect mode button with cva variants
Make selected/hover styles clearer and match RadioGroupItem cursor.
